feat(evolution/infra): regression gate + event retention; drop dead distilled-memory repo
Add real infrastructure and remove only verified-orphan code (net -241 lines).

New:
- ares_evolution regression gate (+unit/wiring tests): rejects candidate
  strategies that regress past performance, complementing the fitness/cost
  penalty; gate_eval tightened.
- event table retention/compaction via a maintenance worker + pg_retention, so
  the Postgres event stream does not grow unbounded.

Removed (RUNTIME #9 distilled_memories ghost — zero production callers):
- storage distilled_memory_repository (+interface) and the builtin
  distilled_memory_tools.go / its memory_tools wiring. The distilled_memories
  DDL is retained deliberately in migrate_storage.go for existing databases,
  so schema compatibility is preserved while the dead repository layer goes away.

Verified: go build ./... and make check (vet, staticcheck, golangci-lint, -race,
closure) pass; closure confirms no orphaned references after the deletion.

+++ internal/runtime/ares_evolution/gate_eval_test.go
@@ -145,13 +145,13 @@ func TestEvalGate_StrictModeRejectsWhenUnconfigured(t *testing.T) {
 // missing-component scenarios.
 func TestEvalGate_SkippedCountIncrements(t *testing.T) {
 	g := NewEvalGate(nil, nil, eval.TestSuite{}, DefaultEvalGateConfig())
-	assert.Equal(t, 0, g.SkippedCount())
+	assert.Equal(t, int64(0), g.SkippedCount())
 
 	g.Check(context.Background(), &mutation.Strategy{}, nil)
-	assert.Equal(t, 1, g.SkippedCount(), "first skip should increment counter")
+	assert.Equal(t, int64(1), g.SkippedCount(), "first skip should increment counter")
 
 	g.Check(context.Background(), &mutation.Strategy{}, nil)
-	assert.Equal(t, 2, g.SkippedCount(), "second skip should increment counter")
+	assert.Equal(t, int64(2), g.SkippedCount(), "second skip should increment counter")
 }
 
 func TestEvalGate_ScoringPath(t *testing.T) {
